Warnings: clean compile with gcc-4.8 at -Wextra -Wno-unused-parameter
* Mostly fills out some ops struct declarations. Reformatted so Emacs
  can help with alignment. Squelch whitespace in diff to see useful
  changes.

* Use struct instead of casting to data pointer (void*) when serializing
  function pointers.

* strcasecmp is in strings.h

+++ src/sys/classes/draw/impls/x/xinit.c
@@ -87,8 +87,8 @@ PetscErrorCode PetscDrawXiDisplayWindow(PetscDraw_X *XiWin,char *label,int x,int
   border_width = 0;
   if (x < 0) x = 0;
   if (y < 0) y = 0;
-  x = ((unsigned int) x + w > wavail) ? wavail - w : x;
-  y = ((unsigned int) y + h > havail) ? havail - h : y;
+  x = ((unsigned int) x + w > wavail) ? wavail - w : (unsigned int)x;
+  y = ((unsigned int) y + h > havail) ? havail - h : (unsigned int)y;
 
   /* We need XCreateWindow since we may need an visual other than the default one */
   XGetWindowAttributes(XiWin->disp,RootWindow(XiWin->disp,XiWin->screen),&in_window_attributes);
